requirements.txt
mangum==0.19.0
psycopg2-binary
python-dotenv
fastapi
uvicorn


# Google AI Veo / GenAI
google-genai==1.43.0
protobuf>=3.20.2,<5

# Optional / Utilities
anyio==4.11.0
annotated-types==0.7.0
blinker==1.9.0
httpx==0.28.1
starlette==0.48.0
pydantic==2.12.3
typing-inspection==0.4.2
